Store parsed content nodes for examples
The content_node key in the example data stores a copy of the parsed
docutils nodes for the example. It turns out that it's easier to use the
parsed content here rather than parse it during the
process_pending_example_nodes() callback where a "state" is not readily
available for parsing reStructuredText.