Generate basic example landing page index.rst
This generates an index.rst landing page for the example gallery
(examples/index.rst). It includes a hidden toctree to each standalone
example page as well as a listing of examples. This listing can be
switched out for a more sophisticated presentation (like cards).

This is fairly basic templating; we might want to switch to Jinja
templates soon.