Add examples/index to demo site
This ensures that the basic tests don't have a "examples/index.rst is
not in a toctree" type errors since the landing page generation for
example pages is automatic.

Maybe it would be better if the example generation was opt-in, or only
activated once an example is actually marked-up in a site?